requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiofiles==0.8.0
    # via quart
aiohttp==3.7.4.post0
    # via
    #   -r requirements.in
    #   twitchio
aspy.refactor-imports==2.2.0
    # via reorder-python-imports
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via
    #   aiohttp
    #   pytest
beautifulsoup4==4.10.0
    # via -r requirements.in
black==21.12b0
    # via -r requirements.in
blinker==1.4
    # via quart
cached-property==1.5.2
    # via aspy.refactor-imports
chardet==4.0.0
    # via aiohttp
click==8.0.3
    # via
    #   black
    #   quart
feedparser==6.0.8
    # via -r requirements.in
h11==0.12.0
    # via
    #   hypercorn
    #   wsproto
h2==4.1.0
    # via hypercorn
hpack==4.0.0
    # via h2
hypercorn==0.12.0
    # via
    #   -r requirements.in
    #   quart
hyperframe==6.0.1
    # via h2
idna==3.3
    # via yarl
iniconfig==1.1.1
    # via pytest
iso8601==1.0.2
    # via twitchio
itsdangerous==2.0.1
    # via quart
jinja2==3.0.3
    # via quart
markupsafe==2.0.1
    # via jinja2
multidict==5.2.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
obs-websocket-py==0.5.3
    # via -r requirements.in
packaging==21.3
    # via pytest
pathspec==0.9.0
    # via black
platformdirs==2.4.0
    # via black
pluggy==1.0.0
    # via pytest
priority==2.0.0
    # via hypercorn
psutil==5.8.0
    # via -r requirements.in
py==1.11.0
    # via pytest
pyparsing==3.0.6
    # via packaging
pytest==6.2.5
    # via -r requirements.in
python-dateutil==2.8.2
    # via -r requirements.in
python-dotenv==0.19.2
    # via -r requirements.in
python-slugify==5.0.2
    # via -r requirements.in
pytz==2021.3
    # via -r requirements.in
pyyaml==6.0
    # via -r requirements.in
quart==0.16.1
    # via -r requirements.in
reorder-python-imports==2.6.0
    # via -r requirements.in
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   obs-websocket-py
    #   python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via
    #   hypercorn
    #   pytest
    #   quart
tomli==1.2.2
    # via black
twitchio==2.1.3
    # via -r requirements.in
typing-extensions==4.0.1
    # via
    #   aiohttp
    #   black
    #   twitchio
websocket-client==1.2.3
    # via obs-websocket-py
werkzeug==2.0.2
    # via quart
wsproto==1.0.0
    # via hypercorn
yarl==1.7.2
    # via aiohttp
